Rename graph_builder.py to graph_construction.py and update references accordingly for clarity.
