+++ fabricpc/core/types.py
@@ -173,7 +173,7 @@ class GraphStructure(NamedTuple):
         n_edges = len(self.edges)
         return f"GraphStructure(nodes={n_nodes}, edges={n_edges})"
 
-    # TODO move to a utility and remove duplicated code in graph_builder.py
+    # TODO move to a utility and remove duplicated code in graph_construction.py
     @staticmethod
     def _topological_sort(
         nodes: Dict[str, Any], edges: Dict[str, "EdgeInfo"]
